bot: address PR #2 review round 3 — pickup lock, single-bot guard, int JSON, sidecar args crash
- LootHandler: take drop lock + setPickedUp like InventoryHandler (dupe race)
- BotSession: volatile bot field; static BOT_IN_USE guard so two sessions
  can't load the same character; despawn on spawn-exception path
- JsonValue: integral number factory — seq/hp no longer serialized as 1.0
- intent_schema: default missing 'args' to {}, reject non-dict (sidecar
  KeyError crash); +2 tests
- ChatHandler: truncate LLM say text to 70 chars

Implemented by OpenCode from review findings; verified: docker javac clean,
JsonSelfCheck OK, sidecar pytest 17 passed.

+++ src/server/bot/json/JsonObject.java
@@ -32,7 +32,7 @@ public class JsonObject extends JsonValue {
 
     public void put(String key, JsonValue value)            { map.put(key, value); }
     public void put(String key, String value)               { map.put(key, JsonValue.string(value)); }
-    public void put(String key, int value)                  { map.put(key, JsonValue.number(value)); }
+    public void put(String key, int value)                  { map.put(key, JsonValue.number((long) value)); }
     public void put(String key, long value)                 { map.put(key, JsonValue.number(value)); }
     public void put(String key, boolean value)              { map.put(key, value ? JsonValue.TRUE : JsonValue.FALSE); }
